+++ src/Inspectors/Character/CharacterInspector.h
@@ -1,6 +1,8 @@
 #ifndef AZERCORE_OPS_CHARACTER_INSPECTOR_H
 #define AZERCORE_OPS_CHARACTER_INSPECTOR_H
 
+#include "Chat.h"
+
 class ChatHandler;
 
 namespace AzerCoreOps
@@ -9,6 +11,7 @@ class CharacterInspector
 {
 public:
     static bool Inspect(ChatHandler* handler);
+    static bool Raid(ChatHandler* handler, Acore::ChatCommands::Tail selection);
     static bool SaveTarget(ChatHandler* handler);
 };
 }